+++ NEWS.md
@@ -1,10 +1,11 @@
 # tbcleanr 0.1.4.0
-#### 14 June 2019
+#### 17 June 2019
 ## Major changes
 
 * `clean_standard_indicators()` added to streamline workflow with Annual TB standard
 indicators Excel data tools.
 * Standard indicators workflow vignette added to package
+* `adhere_classr()` added for EpiInfo and Koch6 data sets
 
 ---
 # tbcleanr 0.1.3.5